Přesunutí consentu z bdsm modulu do nastavení uživatele
