+++ src/Inkluzitron/Modules/UserModule.cs
@@ -8,6 +8,7 @@ using Inkluzitron.Services;
 using Inkluzitron.Utilities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
+using System;
 using System.Threading.Tasks;
 
 namespace Inkluzitron.Modules
@@ -159,6 +160,35 @@ namespace Inkluzitron.Modules
             var user = await DbContext.GetOrCreateUserEntityAsync(Context.User);
             user.Gender = gender;
             await DbContext.SaveChangesAsync();
+            await Context.Message.AddReactionAsync(ReactionSettings.Checkmark);
+        }
+
+        [Command("consent")]
+        [Summary("Vypíše stav souhlasu s používáním obrázkových BDSM příkazů.")]
+        public async Task ShowBdsmConsentAsync([Name("koho")] IUser target = null)
+        {
+            if (target is null)
+                target = Context.User;
+
+            var userEntity = await DbContext.GetOrCreateUserEntityAsync(target);
+            var status = userEntity.HasGivenConsentTo(CommandConsent.BdsmImageCommands);
+            var message = Configuration[status ? "UserModule:ConsentRegistered" : "UserModule:ConsentNotRegistered"];
+            await ReplyAsync(string.Format(message, Format.Sanitize(await UsersService.GetDisplayNameAsync(target))));
+        }
+
+        [Command("consent grant bdsm")]
+        [Summary("Udělí souhlas s používáním obrázkových BDSM příkazů.")]
+        public Task GrantBdsmConsentAsync()
+            => UpdateConsentAsync(c => c | CommandConsent.BdsmImageCommands);
+
+        [Command("consent revoke bdsm")]
+        [Summary("Odvolá souhlas s používáním obrázkových BDSM příkazů.")]
+        public Task RevokeBdsmConsentAsync()
+            => UpdateConsentAsync(c => c & ~CommandConsent.BdsmImageCommands);
+
+        private async Task UpdateConsentAsync(Func<CommandConsent, CommandConsent> consentUpdaterFunc)
+        {
+            await DbContext.UpdateCommandConsentAsync(Context.User, consentUpdaterFunc);
             await Context.Message.AddReactionAsync(ReactionSettings.Checkmark);
         }
     }
