+++ src/Inkluzitron/appsettings.json
@@ -42,20 +42,18 @@
         "StrongTraitThreshold": 0.5,
         "WeakTraitThreshold": 0.2,
         "NoTraitsToReportMessage": "Ve výsledcích testu není k vidění nic spicy. Vanilla it is :slight_smile:",
-        "NoContentToStats": "Vypadá to, že tvému dotazu nelze vyhovět. Zkus prosím upravit kritéria, nebo počkat, až někdo vloží test, kterému půjde vyhovět.",
-        "ConsentRegistered": "{0} souhlasí s používáním obrázkových BDSM příkazů.",
-        "ConsentNotRegistered": "{0} nesouhlasí s používáním obrázkových BDSM příkazů."
+        "NoContentToStats": "Vypadá to, že tvému dotazu nelze vyhovět. Zkus prosím upravit kritéria, nebo počkat, až někdo vloží test, kterému půjde vyhovět."
     },
     "BdsmTraitOperationCheckTranslations": {
         "MissingTargetConsentGendered": [
-            "{0} nezadal/a souhlas s používáním tohoto příkazu (pomocí `$bdsm consent grant`).",
-            "{0} nezadal souhlas s používáním tohoto příkazu (pomocí `$bdsm consent grant`).",
-            "{0} nezadala souhlas s používáním tohoto příkazu (pomocí `$bdsm consent grant`)."
+            "{0} nezadal/a souhlas s používáním tohoto příkazu (pomocí `$consent grant bdsm`).",
+            "{0} nezadal souhlas s používáním tohoto příkazu (pomocí `$consent grant bdsm`).",
+            "{0} nezadala souhlas s používáním tohoto příkazu (pomocí `$consent grant bdsm`)."
         ],
         "MissingUserConsentGendered": [
-            "Nezadal/a jsi souhlas s používáním tohoto příkazu (pomocí `$bdsm consent grant`).",
-            "Nezadal jsi souhlas s používáním tohoto příkazu (pomocí `$bdsm consent grant`).",
-            "Nezadala jsi souhlas s používáním tohoto příkazu (pomocí `$bdsm consent grant`)."
+            "Nezadal/a jsi souhlas s používáním tohoto příkazu (pomocí `$consent grant bdsm`).",
+            "Nezadal jsi souhlas s používáním tohoto příkazu (pomocí `$consent grant bdsm`).",
+            "Nezadala jsi souhlas s používáním tohoto příkazu (pomocí `$consent grant bdsm`)."
         ],
         "MissingTest": "{0} nemá zadán výsledek testu.",
         "Self": "{0} proti sobě nemusí rollovat.",
@@ -106,7 +104,9 @@
     "UserModule": {
         "UserPronounsBotSelf": "jsem já. Říkej mi jak chceš <:peepogiggle:834136184779046973>",
         "UserPronounsMessage": "{0} {1}",
-        "UserNotFoundMessage": "Uživatele {0} se zatím nepodařilo najít v databázi."
+        "UserNotFoundMessage": "Uživatele {0} se zatím nepodařilo najít v databázi.",
+        "ConsentRegistered": "{0} souhlasí s používáním obrázkových BDSM příkazů.",
+        "ConsentNotRegistered": "{0} nesouhlasí s používáním obrázkových BDSM příkazů."
     },
     "PointsKarmaIncrement": 5,
     "UserPointsGraphDays": 30,
